Moved favourite/unfavourite functions into recipecard to reduce props
