+++ app/_components/RecipeContainer.tsx
@@ -1,14 +1,10 @@
 "use client";
 import { useEffect } from "react";
 import { RecipeCard } from "./RecipeCard";
-import { fetchAuthSession } from "aws-amplify/auth";
 
-export function RecipeContainer({
-  recipeResult,
-  onFavouriteRecipe,
-  userId,
-  onUnfavouriteRecipe,
-}) {
+export function RecipeContainer({ recipeResult, userId }) {
+  //mutations
+
   return (
     <div className="grid gap-4 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5">
       {/* {flex space-x-4} */}
@@ -18,13 +14,7 @@ export function RecipeContainer({
         (recipeResult.recipes.length === 0
           ? "No recipes found"
           : recipeResult.recipes.map((recipe) => (
-              <RecipeCard
-                recipeObj={recipe}
-                key={recipe.name}
-                onFavouriteRecipe={onFavouriteRecipe}
-                userId={userId}
-                onUnfavouriteRecipe={onUnfavouriteRecipe}
-              />
+              <RecipeCard recipeObj={recipe} key={recipe.id} userId={userId} />
             )))}
     </div>
   );
